pyproject.toml
[project]
name = "reactor-twin"
version = "1.0.0"
description = "Physics-constrained Neural DEs for chemical reactor digital twins"
authors = [{name = "Tubhyam Karthikeyan", email = "takarthikeyan25@gmail.com"}]
requires-python = ">=3.10"
license = {text = "MIT"}
readme = "README.md"
keywords = [
    "neural-ode", "chemical-reactor", "digital-twin",
    "physics-informed", "surrogate-model", "pinn",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
    "torch>=2.1",
    "torchdiffeq>=0.2.3",
    "numpy>=1.24,<2.0",
    "scipy>=1.11",
    "pandas>=1.5",
    "plotly>=5.18",
    "matplotlib>=3.8",
    "pydantic>=2.5",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dashboard = ["streamlit>=1.28"]
digital_twin = ["scikit-learn>=1.3"]
api = ["fastapi>=0.104", "uvicorn>=0.24", "websockets>=12.0", "prometheus-client>=0.19"]
sde = ["torchsde>=0.2.5"]
cde = ["torchcde>=0.2.5"]
fast = ["torchode>=0.2.0"]
thermo = ["cantera", "coolprop"]
analysis = ["SALib>=1.4"]
discovery = ["pysr>=0.18"]
tracking = ["wandb"]
deploy = ["onnx", "onnxruntime", "boto3>=1.34"]
docs = [
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.24",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "hypothesis>=6.88",
    "ruff>=0.1.6",
    "mypy>=1.7",
    "pre-commit>=3.6",
    "httpx>=0.24",
]
all = ["reactor-twin[dashboard,digital_twin,api,sde,cde,fast,analysis,discovery,tracking,deploy,thermo,docs,dev]"]

[project.scripts]
reactor-twin = "reactor_twin.cli:main"
reactor-twin-dashboard = "reactor_twin.dashboard.app:main"
reactor-twin-api = "reactor_twin.api.server:main"  # Requires [api] extra

[project.urls]
Homepage = "https://github.com/ktubhyam/reactor-twin"
Repository = "https://github.com/ktubhyam/reactor-twin"
Documentation = "https://ktubhyam.github.io/reactor-twin"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "PL"]
ignore = [
    "E501",     # line too long (handled by formatter)
    "PLR0913",  # too many arguments
    "PLR2004",  # magic value comparisons
    "PLR0911",  # too many return statements
    "PLR0912",  # too many branches
    "SIM108",   # ternary operator
    "PLC0415",  # imports not at top level (intentional for optional deps)
    "PLW2901",  # loop variable overwritten
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["PLR2004", "SIM300", "E402"]
"src/reactor_twin/api/*" = ["E402"]
"src/reactor_twin/dashboard/*" = ["E402", "SIM102"]
"examples/*" = ["PLR0915", "B007"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "streamlit", "streamlit.*",
    "torchsde", "torchsde.*",
    "torchcde", "torchcde.*",
    "torchdiffeq", "torchdiffeq.*",
    "torchode", "torchode.*",
    "sklearn", "sklearn.*",
    "plotly", "plotly.*",
    "pandas", "pandas.*",
    "SALib", "SALib.*",
    "scipy", "scipy.*",
    "yaml", "yaml.*",
    "prometheus_client", "prometheus_client.*",
    "onnxruntime", "onnxruntime.*",
    "onnx", "onnx.*",
    "cantera", "cantera.*",
    "coolprop", "coolprop.*",
    "wandb", "wandb.*",
    "pysr", "pysr.*",
    "papermill", "papermill.*",
    "boto3", "boto3.*",
    "botocore", "botocore.*",
    "umap", "umap.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["reactor_twin.dashboard.*"]
ignore_errors = true

[tool.coverage.run]
omit = ["src/reactor_twin/dashboard/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=reactor_twin --cov-report=term-missing"
python_files = "test_*.py"
python_functions = "test_*"
markers = [
    "slow: marks tests as slow (integration tests that cross module boundaries)",
]
